Split admin pipeline modules
